+++ .prettierrc
@@ -1,5 +1,3 @@
 {
-  "trailingComma": "es5",
-  "semi": false,
-  "singleQuote": true
+  "trailingComma": "es5"
 }
