+++ src/components/Layout.js
@@ -22,7 +22,7 @@ const Layout = ({ location, children }) => (
         style={{
           marginLeft: "auto",
           marginRight: "auto",
-          maxWidth: rhythm(24),
+          maxWidth: rhythm(60),
           padding: `${rhythm(1.5)} ${rhythm(3 / 4)}`,
         }}
       >
