+++ src/pages/index.js
@@ -1,21 +1,19 @@
 import React from "react";
-import { Link, graphql } from "gatsby";
+import { graphql } from "gatsby";
 import get from "lodash/get";
 import Helmet from "react-helmet";
 import PropTypes from "prop-types";
 
 import Subtitle from "../components/Subtitle";
 import Layout from "../components/Layout";
-import { rhythm } from "../utils/typography";
 
-class BlogIndex extends React.Component {
+class Index extends React.Component {
   render() {
     const siteTitle = get(this, "props.data.site.siteMetadata.title");
     const siteDescription = get(
       this,
       "props.data.site.siteMetadata.description"
     );
-    const posts = get(this, "props.data.allMarkdownRemark.edges");
 
     return (
       <Layout location={this.props.location} title={siteTitle}>
@@ -25,34 +23,20 @@ class BlogIndex extends React.Component {
           title={siteTitle}
         />
         <Subtitle description={siteDescription} />
-        {posts.map(({ node }) => {
-          const title = get(node, "frontmatter.title") || node.fields.slug;
-          return (
-            <div key={node.fields.slug}>
-              <h3
-                style={{
-                  marginBottom: rhythm(1 / 4),
-                }}
-              >
-                <Link style={{ boxShadow: "none" }} to={node.fields.slug}>
-                  {title}
-                </Link>
-              </h3>
-              <small>{node.frontmatter.date}</small>
-              <p dangerouslySetInnerHTML={{ __html: node.excerpt }} />
-            </div>
-          );
-        })}
+        <p>
+          This is still construction, and will become less horrendous in
+          upcoming weeks. Stay tuned!
+        </p>
       </Layout>
     );
   }
 }
 
-BlogIndex.propTypes = {
+Index.propTypes = {
   location: PropTypes.object,
 };
 
-export default BlogIndex;
+export default Index;
 
 export const pageQuery = graphql`
   query {
@@ -60,20 +44,6 @@ export const pageQuery = graphql`
       siteMetadata {
         title
         description
-      }
-    }
-    allMarkdownRemark(sort: { fields: [frontmatter___date], order: DESC }) {
-      edges {
-        node {
-          excerpt
-          fields {
-            slug
-          }
-          frontmatter {
-            date(formatString: "MMMM DD, YYYY")
-            title
-          }
-        }
       }
     }
   }
